REFACTOR: Clear interfaces folder
