+++ interfaces/index.ts
@@ -1,6 +1,5 @@
-export * from './lesson';
-export * from './settings';
-export * from './class';
-export * from './iscool';
-export * from './timetable';
-export * from './school';
+export * from '@yanshoof/types'
+
+export * from './settings'
+export * from './timetable'
+export * from './school'
