--- interfaces/teacherList.ts
@@ -1,3 +1,0 @@
-export interface ITeacherList {
-  get teachers(): string[];
-}
