+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@yanshoof/iscool": "^1.2.0",
+    "@yanshoof/types": "^1.1.1",
     "axios": "^0.25.0",
     "cookies-next": "^2.0.4",
     "cross-env": "^7.0.3",
